trusted-types/Window-trustedTypes.tentative.html

<!DOCTYPE html>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="support/helper.sub.js"></script>
<body>
<script>
  test(t => {
    let factory = window.trustedTypes;
    assert_true(factory instanceof TrustedTypeFactory);
  }, "factory = window.trustedTypes");
</script>
